dashboard/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.8
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
multidict==6.1.0
packaging==24.1
propcache==0.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.16.0